fix compilation conflicts
